requirements.txt
# FastAPI
FastAPI==0.115.2
uvicorn==0.32.0

# Settings
pydantic-settings==2.6.0
python-multipart==0.0.17

# Data
requests==2.32.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.3
pydantic[email]==2.9.2

# Security
bcrypt==4.2.0
PyJWT==2.9.0
